# doc/api.yaml
openapi: 3.1.0
info:
  title: Wasapp Service API Documentation
  description: |-
    API for interacting with Wasapp, a messaging service.
  version: 1.0.0
servers:
  - url: https://wasapp.online/v1

# ================================================================
...
paths:
...
  /session:
    post:
      tags: ["Login"]
      summary: Log in a user
      description: >-
        If the user does not exist, a new user will be created and both the user identifier and an authentication token will be returned.
        If the user already exists, the user identifier and an authentication token will be returned.
      operationId: doLogin
      requestBody:
        description: User details.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '201':
          description: User created and logged in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request - invalid username (too short or too long or invalid chars).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------- USER SEARCH --------------------
...
  /chats/groups:  # CREATE GROUP
    post:
      tags: ["Chat Creation"]
      summary: Create a group chat
      description: >-
        Create a group chat with the users specified in the request as members. The chat is automatically created by the server with an initial message.
        The autenticated user is automaticaly added to the group chat ea a member.
        The welcome message has the following characteristics: the sender is the user who created the chat, the content is empty, and it has isInitMessage = true.
      operationId: createGroupChat
      security: [ { BearerAuth: [] } ]
      requestBody:
        description: Group details.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroupChatRequest'
      responses:
        '201':
          description: Group chat created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatPreview'
        '400':
          description: Bad request - invalid UserId or invalid GroupName.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Conflict - 1 or more UserID is repeated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{userId}/chats: # CREATE A DIRECT CHAT
    parameters:
      - $ref: '#/components/parameters/UserId'

    post:
      tags: ["Chat Creation"]
      summary: Create a direct chat with a user
      description: >-
        Create a direct chat with the specified user. The chat is automatically created by the server with an initial message.
        The welcome message has the following characteristics: the sender is the user who created the chat, the content is empty, and it has isInitMessage = true.
      operationId: createDirectChat
      security: [ { BearerAuth: [] } ]
      responses:
        '201':
          description: Direct chat created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatPreview'
        '401':
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: You already have a direct chat with this user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------- USER AVATAR IMAGE --------------------
...
  /chats/{chatId}/name:
    parameters:
      - $ref: '#/components/parameters/ChatId'

    patch:
      tags: ["Group Operations"]
      summary: Change the name of the specified group
      description: >-
        Changes the name of the group identified by the specified chat ID,
        using the new name provided in the request body.
        If the given identifier is not associated with a group chat, an error will be returned.
      operationId: setGroupName
      security: [ { BearerAuth: [] } ]
      requestBody:
        description: The new name for the group.
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/PatchGroupNameRequest'
      responses:
        "200":
          description: Group name updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatName'
        "400":
          description: Bad request - invalid group name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden - you are not authorized to change this group name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Group not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: This is not a group chat, only the names of group chats can be set/changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -------------------- GROUP AVATAR IMAGE --------------------
...
  /chats/{chatId}/members/me:
    parameters:
      - $ref: '#/components/parameters/ChatId'

    delete:
      tags: ["Group Operations"]
      summary: Leave the group
      description: Removes the logged-in user from the group identified by the specified chat ID.
      security: [ { BearerAuth: [] } ]
      operationId: leaveGroup
      responses:
        "204":
          description: User left the group successfully. No content is returned.
        "401":
          description: Authentication required or failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden - you are not a member of this group or this is not a group chat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Group not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: This is not a group chat, only members of group chats can leave the group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      description: Unique identifier for the user (UUID).
      required: true
      schema: { $ref: '#/components/schemas/UserId' }
    ChatId:
      name: chatId
      in: path
      description: Unique identifier for the chat (UUID).
      required: true
      schema: { $ref: '#/components/schemas/ChatId' }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT"

  # ================================================================
  schemas:
...
    BaseId:
      type: string
      format: uuid
      readOnly: true
      description: Base UUID identifier type, used only as a parent for specific ID schemas
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
      example: "550e8400-e29b-41d4-a716-446655440000"
    BaseName:
      description: >-
        The name must be 3 to 24 characters long.
        Allowed characters are letters, numbers, spaces, underscores (_), and hyphens (-).
        The name needs at least one alpha-numeric character.
      type: string
      pattern: "^(?=.*[A-Za-z0-9_-])[A-Za-z0-9 _-]{3,24}$"
      minLength: 3
      maxLength: 24
    BaseDateTime:
      type: string
      format: date-time
      description: ISO 8601 timestamp
      example: "2025-11-11T09:59:00Z"
      minLength: 20
      maxLength: 30

    # -------------------- BASE IMAGE --------------------
...
    ImageUrl:
      type: string
      format: url
      description: >-
        URL to access a message image. Image format is indicated by the HTTP Content-Type header upon retrieval.
      example: "https://wasapp.online/v1/chats/550e8400-e29b-41d4-a716-446655440000/images/660e8400-e29b-41d4-a716-446655440000"
...
    UserId:
      allOf:
        - $ref: '#/components/schemas/BaseId'
        - description: Unique identifier assigned to the user by the server.
    UserName:
      allOf:
        - $ref: '#/components/schemas/BaseName'
        - description: Unique username visible to other users.
        - example: Feldspar

    # -------------------- LOGIN --------------------
    LoginRequest:
      type: object
      description: Request payload containing the username to log in or create a user.
      required: [userName]
      properties:
        userName: { $ref: '#/components/schemas/UserName' }
    BearerToken:
      type: string
      description: Token used to fulfill the Bearer Authentication standard.
      pattern: ^Bearer\s[A-Za-z0-9-._~+/]+={0,2}$
      example: "Bearer JhdkEN7QWidlYDjRTH8I~n99YJfqCC=="

    # -------------------- CHAT --------------------
    ChatId:
      allOf:
        - $ref: '#/components/schemas/BaseId'
        - description: Unique identifier assigned to the chat by the server.
    ChatName:
      allOf:
        - $ref: '#/components/schemas/BaseName'
        - description: Name of the chat
        - example: Marco34
    ChatPreview:
      type: object
      description: >-
        Represents a chat. Standard chats include two users;
        group chats can be created with the group flag.
      required: [id, displayName, lastMessage]
      properties:
        id:           { $ref: '#/components/schemas/ChatId'          }
        displayName:  { $ref: '#/components/schemas/ChatDisplayName' }
        isGroupChat:  { $ref: '#/components/schemas/IsGroupChat'     }
        lastMessage:  { $ref: '#/components/schemas/ChatLastMessage' }
    ChatDisplayName:
      allOf:
        - $ref: '#/components/schemas/ChatName'
        - description: |-
            The name displayed in the chat preview.
              - For 1-on-1 chats: the name of the other user is displayed.
              - For group chats: the group name is displayed.
    IsGroupChat:
      type: boolean
      description: >-
        Indicates if it's a group chat. Used by the client to determine when to display additional group info.
        For example, if is a group chat, the client can show in the last message preview the sender's name.
      readOnly: true
    ChatLastMessage:
      type: object
      description: |-
        Preview of the last message sent in the chat. It can be composed by a text or by an image or both in this case the text is used as caption.
        - If the 'isDeleted' flag is true, the message content will not be returned by the server, and the client should render it as a deleted message.
        - If the 'isInitMessage' flag is true, the message content will not be returned by the server, and the client should render it as a "welcome" message.
      required: [status, sendTime, senderName, isDeleted, isInitMessage]
      properties:
        status:        { $ref: '#/components/schemas/MessageStatus'        }
        sendTime:      { $ref: '#/components/schemas/BaseDateTime'         }
        senderName:    { $ref: '#/components/schemas/UserName'             }
        isDeleted:     { $ref: '#/components/schemas/MessageIsDeletedFlag' }
        isInitMessage: { $ref: '#/components/schemas/IsInitMessageFlag'    }
        content:       { $ref: '#/components/schemas/MessageContent'       } #optional
...
    ChatMembersListIdOnly:
      type: array
      description: List of users in a chat, ordered alphabetically by username.
      minItems: 1
      maxItems: 1024 #max number of allowed members in a group chat
      items:
        $ref: '#/components/schemas/UserId'

    # -------------------- MESSAGE --------------------
...
    MessageContent:
      description: >-
        Generic message content: only text, only image, or both
        (in this case, the text is used as a caption for the image).
      oneOf:
        - type: object
          required: [text, msgImageUrl]
          properties:
            text:        { $ref: '#/components/schemas/MessageText' }
            msgImageUrl: { $ref: '#/components/schemas/ImageUrl'    }
        #--------------------------
        - type: object
          required: [text]
          properties:
            text: { $ref: '#/components/schemas/MessageText' }
        #--------------------------
        - type: object
          required: [msgImageUrl]
          properties:
            msgImageUrl: { $ref: '#/components/schemas/ImageUrl' }
    MessageText:
      type: string
      description: >-
        The message text. Must contain at least one non-whitespace character.
        Cannot be made up only of spaces or other whitespace characters.
      minLength: 1
      maxLength: 3000
      pattern: '^(?!\s*$).+'
      example: "Hello World!!"
    MessageIsDeletedFlag:
      type: boolean
      description: >-
        Indicates whether the message content has been deleted.
        If true, the message content is not returned by the server,
        and the client should display this message as deleted.
    IsInitMessageFlag:
      type: boolean
      description: >-
        Indicates whether the message is the default message generated by the server when a new chat is created.
        When this flag is true, the message content should be empty and ignored by the client.
        The client should render the message as a welcome message like "This group was created by {senderOfTheMessage}"
        or "{senderOfTheMessage} started a direct chat with you". A message with the IsInitMessageFlag set to true can't be deleted
    MessageStatus:
      type: string
      enum: [delivered, received, read]
      description: |-
        Message read status:
          - 'delivered' means the message has been delivered to the server (default)
          - 'received' means the message has been received by all recipients (all participants except the sender)
          - 'read' means the message has been read by all recipients (all participants except the sender)
        This info is only useful when the requesting user sent the message, so the app can show in the preview of the chat if others have received/read your message.
...
    PatchGroupNameRequest:
      type: object
      description: Request body for changing the name of the specified group.
      properties:
        groupName:
          $ref: '#/components/schemas/ChatName'
    NewGroupChatRequest:
      type: object
      description: >-
        Details necessary for the creation of the new group chat.
        The list should not include the user who is creating the chat,
        since they are implicitly added to the users of a group.
      required: [groupName, membersList]
      properties:
        groupName:   { $ref: '#/components/schemas/ChatName'        }
        membersList: { $ref: '#/components/schemas/ChatMembersListIdOnly' }
...
    Error:
      type: object
      description: Default error structure.
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code for frontend logic.
          enum:
            - CHAT_NOT_FOUND
            - USER_NOT_FOUND
            - ALREADY_IN_GROUP
            - NOT_A_GROUP_CHAT
            - FORBIDDEN_NOT_MEMBER
            - INVALID_INPUT
            - INTERNAL_SERVER_ERROR
            - UNAUTHORIZED
            - USERNAME_TAKEN
            - CHAT_ALREADY_EXISTS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
          example: USER_NOT_FOUND
        message:
          type: string
          description: Human-readable error message.
          example: The specified user ID does not exist.
    AuthResponse:
      type: object
      description: Authentication result containing the user identifier and bearer token.
      readOnly: true
      required: [identifier, token]
      properties:
        identifier: { $ref: '#/components/schemas/UserId'      }
        token:      { $ref: '#/components/schemas/BearerToken' }
...